requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.6.3
bs4==0.0.1
builtwith==1.3.3
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
future==0.16.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jieba==0.39
lxml==4.2.5
MechanicalSoup==0.11.0
parsel==1.5.0
Pillow==5.2.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pymongo==3.7.1
PyMySQL==0.9.2
pyOpenSSL==18.0.0
pytesseract==0.2.4
python-whois==0.7.0
pywin32==223
queuelib==1.5.0
requests==2.19.1
Scrapy==1.5.1
selenium==3.14.1
service-identity==17.0.0
six==1.11.0
tesseract==0.1.3
Twisted==18.7.0
urllib3==1.23
w3lib==1.19.0
zope.interface==4.5.0
